+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import { ProjectSection } from "@/components/ProjectSection/ProjectSection";
 import HubTemplate from "@/templates/HubTemplate";
 import { getContent } from "@/utils/functions";
 import { SectionHeader } from "@/utils/interfaces";
-import { AboutSection } from "@/components/About/About";
+import { AboutSection } from "@/components/AboutSection/AboutSection";
 import { RecentSection } from "@/components/RecentSection/RecentSection";
 import PreviewSection from "@/components/PreviewSection/PreviewSection";
 import MainBanner from "@/components/BannerImage/BannerImage";
